requirements.txt
fastapi[standard]
pg8000
python-dotenv
requests